Add flow chart
