--- a-tour-of-go/basics.go
@@ -1,91 +1,0 @@
-package main
-
-import (
-	"fmt"
-	"math"
-	"math/cmplx"
-	"math/rand"
-)
-
-var c, python, java bool = true, false, false
-
-var (
-	ToBe   bool       = false
-	MaxInt uint64     = 1<<64 - 1
-	z      complex128 = cmplx.Sqrt(-5 + 12i)
-)
-
-func main() {
-	fmt.Println("My favorite number is", rand.Intn(10))
-
-	fmt.Println("Now you have %g problems.", math.Sqrt(7))
-
-	fmt.Println(math.Pi) // Exported names start with Capital letter.
-
-	fmt.Println(add(42, 13))
-
-	a, b := swap("hello", "world")
-	fmt.Println(a, b)
-
-	fmt.Println(split(17))
-
-	var i, j = 1, "no"
-	k := "same-as-var"
-	fmt.Println(i, j, k, c, python, java)
-
-	const f = "%T(%v)\n"
-	fmt.Printf(f, ToBe, ToBe)
-	fmt.Printf(f, MaxInt, MaxInt)
-	fmt.Printf(f, z, z)
-
-	zero_values()
-	type_conversion()
-	testConst()
-}
-
-func add(x, y int) int {
-	return x + y
-}
-
-func swap(x, y string) (string, string) {
-	return y, x
-}
-
-func split(sum int) (x, y int) {
-	x = sum * 4 / 9
-	y = sum - x
-	return
-}
-
-func zero_values() {
-	var i int
-	var f float64
-	var b bool
-	var s string
-	fmt.Printf("%v %v %v %q\n", i, f, b, s)
-}
-
-func type_conversion() {
-	var x, y int = 3, 4
-	var f float64 = math.Sqrt(float64(x*x + y*y))
-	var z uint = uint(f)
-	fmt.Println(x, y, z)
-}
-
-const (
-	Big   = 1 << 100
-	Small = Big >> 99
-)
-
-func needInt(x int) int { return x*10 + 1 }
-func needFloat(x float64) float64 {
-	return x * 0.1
-}
-
-func testConst() {
-	fmt.Println(needInt(Small))
-	fmt.Println(needFloat(Small))
-	fmt.Println(needFloat(Big))
-
-	// fmt.Println(needInt(Big)) // over flows
-}
